+++ CHANGELOG.md
@@ -2,3 +2,11 @@
 
 ## 1.0.0
 - Initial release
+
+## 1.1.0
+- Change the cursor for the buttons inside the sidebar (add folder)
+- Remove the [unsupported] from the App title
+
+## 1.1.1
+- Updated selectors for changes in TabBar classes
+- Added custom message when the extension is updated